+++ test_scripts/rust/test_ag2/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "temp_rust_test"
+version = "0.1.1"
+edition = "2021"
+
+[dependencies]
+# Assuming you're using the runagent Rust SDK
+runagent = { path = "/home/riamdriad5/runagent/runagent/runagent-rust/runagent" }
+
+# Required dependencies
+tokio = { version = "1.0", features = ["full"] }
+serde_json = "1.0"
+anyhow = "1.0"
+futures = "0.3"
